+++ test/HCurveDao.test.js
@@ -2,25 +2,18 @@ const {
   BN,
   ether,
   expectRevert,
-  time,
   constants,
 } = require('@openzeppelin/test-helpers');
 const { MAX_UINT256 } = constants;
 const { expect } = require('chai');
 const abi = require('ethereumjs-abi');
 const utils = web3.utils;
-const { increase, duration } = time;
-const {
-  CRV_TOKEN,
-  CURVE_AAVECRV,
-  CURVE_AAVECRV_PROVIDER,
-  CURVE_AAVE_GAUGE,
-} = require('./utils/constants');
+const { CURVE_AAVECRV, CURVE_AAVE_GAUGE } = require('./utils/constants');
 const {
   evmRevert,
   evmSnapshot,
   profileGas,
-  getHandlerReturn,
+  tokenProviderCurveGauge,
 } = require('./utils/utils');
 
 const Proxy = artifacts.require('ProxyMock');
@@ -34,11 +27,14 @@ contract('Curve DAO', function([_, user]) {
   let id;
   // Wait for the gaude to be ready
   const token0Address = CURVE_AAVECRV;
-  const token0Provider = CURVE_AAVECRV_PROVIDER;
   const gauge0Address = CURVE_AAVE_GAUGE;
   const gauge0Amount = ether('0.1');
 
+  let token0Provider;
+
   before(async function() {
+    token0Provider = await tokenProviderCurveGauge(token0Address);
+
     this.registry = await Registry.new();
     this.hCurveDao = await HCurveDao.new();
     await this.registry.register(
@@ -48,11 +44,6 @@ contract('Curve DAO', function([_, user]) {
     this.token0 = await IToken.at(token0Address);
     this.gauge0 = await ILiquidityGauge.at(gauge0Address);
     this.proxy = await Proxy.new(this.registry.address);
-
-    await hre.network.provider.request({
-      method: 'hardhat_impersonateAccount',
-      params: [token0Provider],
-    });
   });
 
   beforeEach(async function() {