+++ test/HFunds.test.js
@@ -5,12 +5,10 @@ const {
   ether,
   expectEvent,
   expectRevert,
-  time,
   send,
 } = require('@openzeppelin/test-helpers');
 const { ZERO_ADDRESS, MAX_UINT256 } = constants;
 const { tracker } = balance;
-const { latest } = time;
 const abi = require('ethereumjs-abi');
 const utils = web3.utils;
 
@@ -19,12 +17,8 @@ const { expect } = require('chai');
 const {
   MATIC_TOKEN,
   DAI_TOKEN,
-  DAI_PROVIDER,
   BAT_TOKEN,
-  BAT_PROVIDER,
   USDT_TOKEN,
-  USDT_PROVIDER,
-  MATIC_PROVIDER_CONTRACT,
   NATIVE_TOKEN,
 } = require('./utils/constants');
 const {
@@ -32,6 +26,9 @@ const {
   evmSnapshot,
   profileGas,
   getHandlerReturn,
+  maticProviderWmatic,
+  tokenProviderQuick,
+  tokenProviderSushi,
 } = require('./utils/utils');
 
 const HFunds = artifacts.require('HFunds');
@@ -44,10 +41,20 @@ contract('Funds', function([_, user, someone]) {
   let id;
   let balanceUser;
   let balanceProxy;
-  const tokenAddresses = [DAI_TOKEN, BAT_TOKEN];
-  const providerAddresses = [DAI_PROVIDER, BAT_PROVIDER];
+  const token0Address = DAI_TOKEN;
+  const token1Address = BAT_TOKEN;
+
+  let provider0Address;
+  let provider1Address;
+  let usdtProviderAddress;
+  let maticProviderAddress;
 
   before(async function() {
+    provider0Address = await tokenProviderQuick(token0Address);
+    provider1Address = await tokenProviderSushi(token1Address);
+    usdtProviderAddress = await tokenProviderQuick(USDT_TOKEN);
+    maticProviderAddress = await maticProviderWmatic();
+
     this.registry = await Registry.new();
     this.proxy = await Proxy.new(this.registry.address);
     this.hFunds = await HFunds.new();
@@ -55,23 +62,6 @@ contract('Funds', function([_, user, someone]) {
       this.hFunds.address,
       utils.asciiToHex('Funds')
     );
-
-    await hre.network.provider.request({
-      method: 'hardhat_impersonateAccount',
-      params: [providerAddresses[0]],
-    });
-    await hre.network.provider.request({
-      method: 'hardhat_impersonateAccount',
-      params: [providerAddresses[1]],
-    });
-    await hre.network.provider.request({
-      method: 'hardhat_impersonateAccount',
-      params: [MATIC_PROVIDER_CONTRACT],
-    });
-    await hre.network.provider.request({
-      method: 'hardhat_impersonateAccount',
-      params: [USDT_PROVIDER],
-    });
   });
 
   beforeEach(async function() {
@@ -84,8 +74,8 @@ contract('Funds', function([_, user, someone]) {
 
   describe('update tokens', function() {
     before(async function() {
-      this.token0 = await IToken.at(tokenAddresses[0]);
-      this.token1 = await IToken.at(tokenAddresses[1]);
+      this.token0 = await IToken.at(token0Address);
+      this.token1 = await IToken.at(token1Address);
       balanceUser = await tracker(user);
       balanceProxy = await tracker(this.proxy.address);
     });
@@ -96,10 +86,10 @@ contract('Funds', function([_, user, someone]) {
       const to = this.hFunds.address;
       const data = abi.simpleEncode('updateTokens(address[])', token);
       await this.token0.transfer(this.proxy.address, value[0], {
-        from: providerAddresses[0],
+        from: provider0Address,
       });
       await this.token1.transfer(this.proxy.address, value[1], {
-        from: providerAddresses[1],
+        from: provider1Address,
       });
 
       const receipt = await this.proxy.execMock(to, data, {
@@ -133,10 +123,10 @@ contract('Funds', function([_, user, someone]) {
       const data = abi.simpleEncode('updateTokens(address[])', token);
       // Transfer tokens to proxy first
       await this.token0.transfer(this.proxy.address, value[0], {
-        from: providerAddresses[0],
+        from: provider0Address,
       });
       // Proxy does not allow transfer ether from EOA so we use provider contract
-      await send.ether(MATIC_PROVIDER_CONTRACT, this.proxy.address, value[1]);
+      await send.ether(maticProviderAddress, this.proxy.address, value[1]);
       await balanceUser.get();
 
       const receipt = await this.proxy.execMock(to, data, {
@@ -171,10 +161,10 @@ contract('Funds', function([_, user, someone]) {
       const data = abi.simpleEncode('updateTokens(address[])', token);
       // Transfer tokens to proxy first
       await this.token0.transfer(this.proxy.address, value[0], {
-        from: providerAddresses[0],
+        from: provider0Address,
       });
       // Proxy does not allow transfer ether from EOA so we use provider contract
-      await send.ether(MATIC_PROVIDER_CONTRACT, this.proxy.address, value[1]);
+      await send.ether(maticProviderAddress, this.proxy.address, value[1]);
       await balanceUser.get();
 
       const receipt = await this.proxy.execMock(to, data, {
@@ -219,7 +209,7 @@ contract('Funds', function([_, user, someone]) {
   describe('inject', function() {
     describe('single token', function() {
       before(async function() {
-        this.token0 = await IToken.at(tokenAddresses[0]);
+        this.token0 = await IToken.at(token0Address);
         this.usdt = await IUsdt.at(USDT_TOKEN);
       });
 
@@ -233,7 +223,7 @@ contract('Funds', function([_, user, someone]) {
           value
         );
         await this.token0.transfer(user, value[0], {
-          from: providerAddresses[0],
+          from: provider0Address,
         });
         await this.token0.approve(this.proxy.address, value[0], { from: user });
 
@@ -267,7 +257,7 @@ contract('Funds', function([_, user, someone]) {
           value
         );
         await this.usdt.transfer(user, value[0], {
-          from: USDT_PROVIDER,
+          from: usdtProviderAddress,
         });
         await this.usdt.approve(this.proxy.address, value[0], { from: user });
 
@@ -313,8 +303,8 @@ contract('Funds', function([_, user, someone]) {
 
     describe('multiple tokens', function() {
       before(async function() {
-        this.token0 = await IToken.at(tokenAddresses[0]);
-        this.token1 = await IToken.at(tokenAddresses[1]);
+        this.token0 = await IToken.at(token0Address);
+        this.token1 = await IToken.at(token1Address);
       });
 
       it('normal', async function() {
@@ -327,11 +317,11 @@ contract('Funds', function([_, user, someone]) {
           value
         );
         await this.token0.transfer(user, value[0], {
-          from: providerAddresses[0],
+          from: provider0Address,
         });
         await this.token0.approve(this.proxy.address, value[0], { from: user });
         await this.token1.transfer(user, value[1], {
-          from: providerAddresses[1],
+          from: provider1Address,
         });
         await this.token1.approve(this.proxy.address, value[1], { from: user });
 
@@ -377,7 +367,7 @@ contract('Funds', function([_, user, someone]) {
           value
         );
         await this.token0.transfer(user, value[0], {
-          from: providerAddresses[0],
+          from: provider0Address,
         });
         await this.token0.approve(this.proxy.address, value[0], { from: user });
 
@@ -394,14 +384,14 @@ contract('Funds', function([_, user, someone]) {
 
   describe('send', function() {
     before(async function() {
-      this.token = await IToken.at(tokenAddresses[0]);
+      this.token = await IToken.at(token0Address);
       this.usdt = await IUsdt.at(USDT_TOKEN);
     });
 
     describe('token', function() {
       it('normal', async function() {
         const token = this.token.address;
-        const providerAddress = providerAddresses[0];
+        const providerAddress = provider0Address;
         const value = ether('100');
         const receiver = someone;
         const to = this.hFunds.address;
@@ -433,7 +423,7 @@ contract('Funds', function([_, user, someone]) {
 
       it('USDT', async function() {
         const token = this.usdt.address;
-        const providerAddress = USDT_PROVIDER;
+        const providerAddress = usdtProviderAddress;
         const value = new BN('1000000');
         const receiver = someone;
         const to = this.hFunds.address;
@@ -466,7 +456,7 @@ contract('Funds', function([_, user, someone]) {
 
       it('maximum', async function() {
         const token = this.token.address;
-        const providerAddress = providerAddresses[0];
+        const providerAddress = provider0Address;
         const value = ether('10');
         const receiver = someone;
         const to = this.hFunds.address;
@@ -498,7 +488,7 @@ contract('Funds', function([_, user, someone]) {
 
       it('send 0 token', async function() {
         const token = this.token.address;
-        const providerAddress = providerAddresses[0];
+        const providerAddress = provider0Address;
         const value = ether('0');
         const receiver = someone;
         const to = this.hFunds.address;
@@ -535,7 +525,7 @@ contract('Funds', function([_, user, someone]) {
 
       it('insufficient token', async function() {
         const token = this.token.address;
-        const providerAddress = providerAddresses[0];
+        const providerAddress = provider0Address;
         const value = ether('100');
         const receiver = someone;
         const to = this.hFunds.address;
@@ -646,7 +636,7 @@ contract('Funds', function([_, user, someone]) {
     describe('multiple tokens', function() {
       before(async function() {
         this.token0 = this.usdt;
-        this.token1 = await IToken.at(tokenAddresses[1]);
+        this.token1 = await IToken.at(token1Address);
       });
 
       it('multiple tokens', async function() {
@@ -662,10 +652,10 @@ contract('Funds', function([_, user, someone]) {
         );
 
         await this.token0.transfer(this.proxy.address, value[0], {
-          from: USDT_PROVIDER,
+          from: usdtProviderAddress,
         });
         await this.token1.transfer(this.proxy.address, value[1], {
-          from: providerAddresses[1],
+          from: provider1Address,
         });
 
         await this.proxy.updateTokenMock(this.token0.address);
@@ -715,7 +705,7 @@ contract('Funds', function([_, user, someone]) {
         );
 
         await this.token1.transfer(this.proxy.address, value[1], {
-          from: providerAddresses[1],
+          from: provider1Address,
         });
 
         await this.proxy.updateTokenMock(this.token1.address);
@@ -755,7 +745,7 @@ contract('Funds', function([_, user, someone]) {
         );
 
         await this.token1.transfer(this.proxy.address, value[1], {
-          from: providerAddresses[1],
+          from: provider1Address,
         });
 
         await this.proxy.updateTokenMock(this.token1.address);
@@ -795,7 +785,7 @@ contract('Funds', function([_, user, someone]) {
         );
 
         await this.token1.transfer(this.proxy.address, value[1], {
-          from: providerAddresses[1],
+          from: provider1Address,
         });
 
         await this.proxy.updateTokenMock(this.token1.address);
@@ -839,7 +829,7 @@ contract('Funds', function([_, user, someone]) {
         );
 
         await this.token1.transfer(this.proxy.address, value[1], {
-          from: providerAddresses[1],
+          from: provider1Address,
         });
 
         await this.proxy.updateTokenMock(this.token1.address);
@@ -868,7 +858,7 @@ contract('Funds', function([_, user, someone]) {
         );
 
         await this.token0.transfer(this.proxy.address, value[0], {
-          from: USDT_PROVIDER,
+          from: usdtProviderAddress,
         });
 
         await this.proxy.updateTokenMock(this.token0.address);
@@ -885,14 +875,14 @@ contract('Funds', function([_, user, someone]) {
 
   describe('get balance', function() {
     before(async function() {
-      this.token = await IToken.at(tokenAddresses[0]);
+      this.token = await IToken.at(token0Address);
       this.usdt = await IUsdt.at(USDT_TOKEN);
     });
     describe('Ether', async function() {
       it('normal', async function() {
         const token = constants.ZERO_ADDRESS;
         const value = ether('1');
-        const providerAddress = providerAddresses[0];
+        const providerAddress = provider0Address;
         const to = this.hFunds.address;
         const data = abi.simpleEncode(
           'getBalance(address):(uint256)',
@@ -917,7 +907,7 @@ contract('Funds', function([_, user, someone]) {
         it('normal', async function() {
           const token = this.token.address;
           const value = ether('1');
-          const providerAddress = providerAddresses[0];
+          const providerAddress = provider0Address;
           const to = this.hFunds.address;
           const data = abi.simpleEncode('getBalance(address):(uint256)', token);
           await this.token.transfer(this.proxy.address, value, {
@@ -942,8 +932,8 @@ contract('Funds', function([_, user, someone]) {
 
   describe('check slippage', function() {
     before(async function() {
-      this.token0 = await IToken.at(tokenAddresses[0]);
-      this.token1 = await IToken.at(tokenAddresses[1]);
+      this.token0 = await IToken.at(token0Address);
+      this.token1 = await IToken.at(token1Address);
     });
 
     it('normal', async function() {
@@ -957,11 +947,11 @@ contract('Funds', function([_, user, someone]) {
       );
 
       await this.token0.transfer(this.proxy.address, value[0], {
-        from: providerAddresses[0],
+        from: provider0Address,
       });
 
       await this.token1.transfer(this.proxy.address, value[1], {
-        from: providerAddresses[1],
+        from: provider1Address,
       });
 
       const receipt = await this.proxy.execMock(to, data, {
@@ -983,11 +973,11 @@ contract('Funds', function([_, user, someone]) {
       );
 
       await this.token0.transfer(this.proxy.address, value[0], {
-        from: providerAddresses[0],
+        from: provider0Address,
       });
 
       await this.token1.transfer(this.proxy.address, value[1], {
-        from: providerAddresses[1],
+        from: provider1Address,
       });
       revertValue = ether('1');
 
@@ -1012,11 +1002,11 @@ contract('Funds', function([_, user, someone]) {
 
       revertValue = ether('1');
       await this.token0.transfer(this.proxy.address, revertValue, {
-        from: providerAddresses[0],
+        from: provider0Address,
       });
 
       await this.token1.transfer(this.proxy.address, value[1], {
-        from: providerAddresses[1],
+        from: provider1Address,
       });
 
       await expectRevert(
@@ -1039,7 +1029,7 @@ contract('Funds', function([_, user, someone]) {
       );
 
       await this.token0.transfer(this.proxy.address, value[0], {
-        from: providerAddresses[0],
+        from: provider0Address,
       });
 
       await expectRevert(