+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Migrate developing environment from Truffle to Hardhat.
+- Refine token provider for tests.
 
 ## [1.9.0] - 2021-10-01